Update: 文節解析を統合＋改善点ドキュメントの追加
